requirements.txt
attrs==19.3.0
aws-cdk.assets==1.17.1
aws-cdk.aws-certificatemanager==1.17.1
aws-cdk.aws-cloudformation==1.17.1
aws-cdk.aws-cloudwatch==1.17.1
aws-cdk.aws-ec2==1.17.1
aws-cdk.aws-events==1.17.1
aws-cdk.aws-iam==1.17.1
aws-cdk.aws-kms==1.17.1
aws-cdk.aws-lambda==1.17.1
aws-cdk.aws-logs==1.17.1
aws-cdk.aws-route53==1.17.1
aws-cdk.aws-s3==1.17.1
aws-cdk.aws-s3-assets==1.17.1
aws-cdk.aws-sns==1.17.1
aws-cdk.aws-sqs==1.17.1
aws-cdk.aws-ssm==1.17.1
aws-cdk.core==1.17.1
aws-cdk.cx-api==1.17.1
aws-cdk.region-info==1.17.1
-e git+git@github.com:infonas/cdk-binnydemohostnet.git@c5206074edbdb12ec50c0b02fd4cf3f15824f1d1#egg=binnydemohostnet
boto3==1.10.16
botocore==1.13.16
cattrs==0.9.0
docutils==0.15.2
importlib-resources==1.0.2
jmespath==0.9.4
jsii==0.20.7
mypy-extensions==0.4.3
pkg-resources==0.0.0
publication==0.0.3
python-dateutil==2.8.1
s3transfer==0.2.1
six==1.13.0
typing-extensions==3.7.4.1
urllib3==1.25.7
